Started using the LuaLogging library. In the future, I should probably include a list of libraries that I am using here. For now, just be aware that I am using Lua For Windows, which has about 30 libraries included by default (things like LuaSQL, LuaSocket, LuaUnit, LuaLogging, and all sorts of good stuff).
Also, as it stands, I am using 2 loggers to accomplish the task of logging
Game events, because there seems to only be 1 option for each logger: have a
logger that outputs to a file, or have a logger that outputs to the console.
As such, I'm using one for each. I also have one logger setup to log client
input (which only goes to a file).