Added Mob object, which replaced the Creature object. Later on this'll probably inherit from some kind of MapObject parent. For the time being, it'll do.
The Player interactions are mostly fleshed out now. I just have one huge
problem.

The method in which I get the context of the "packages," the ones which
contain a singleton which I access throughout the game, is not gonna work. I
have the Nanny requiring access to the Game, and the Game requiring access
to the Nanny. And to access either, I have to require() them.

I may have to do something like just have the Game be a globally accessible
entity to start with. We'll see.