Just some more restructuring, so things are more consistent. Map stuff is mostly done. Now that the underlying framework that governs how MapObjects, MapTiles, and Maps interact is there, I can start working on implementing the concept of "directions" and directional movement.
The relationship between a MapObject and its location (and the map it
inhabits, as such) has been more solidified, so that they are mutual and
binding. When one changes, so does the other.

Same with a Player and its Mob. Player.mob and Mob.player are both mutual
associations that must be shared across both, so when one changes, so does
the other.