Added information to README to explain basic features of the project, just for the finer details.
The first argument to main.lua, if given, will be used as the port to host
on. Otherwise, it will use Game.defaultPort. Added a 1/10th second delay
between Game.update() cycles in main.lua's game loop.

Events, when firing, now use the timestamp in which they fired as the basis
for the next destination in scenarios where the Event will repeat. This is
different from the original setup, which used the previous destination as
the basis for the next destination. The previous setup essentially meant
that, if the starting destination was set to 0 or some time before
os.clock(), there could arise a scenario where the event can't catch up to
os.clock().

Updated version numbers and documentation a bit.