Reorganized Game.lua.
Reorganizing require()s so that instead of requiring packages for clones that you interact with in another package, you only require packages for clones that you make copies of, as this is the only time you actually need access to the pure clone.

Other times you might need access to pure clones is when using things like
Cloneable.isCloneOf(), for determining if something is a clone of say,
Event, or Mob.