Adding some more useful information to Classifications.
Added the basic documentation comments for the new attribute getters.

Removed safeiPair and safePair all together. I was using it for the Scheduler and to poll players for input. Created a copy of a table 100 times a second is pretty stupid. It might be useful some time, but for now I just threw it out.

When logging into an existing character, the existing mob data is not assigned to the Player. It is instead put in the Player's nanny table, which should be used for all persistent data used for handling a new connection.

Moved matchKeywords to the string extension file.